backend/requirements.txt
fastapi
uvicorn
pydantic
requests
httpx
python-dotenv